chore: make Target(version) u32 instead of i32, feature gate build_db_url to url databases

+++ refinery_core/src/traits/async.rs
@@ -27,7 +27,7 @@ async fn migrate<T: AsyncTransaction>(
 ) -> Result<(), Error> {
     for migration in migrations.iter() {
         if let Target::Version(input_target) = target {
-            if input_target < migration.version {
+            if (input_target as i32) < migration.version {
                 log::info!("stoping at migration: {}, due to user option", input_target);
                 break;
             }
@@ -54,7 +54,7 @@ async fn migrate_grouped<T: AsyncTransaction>(
     let mut display_migrations = Vec::new();
     for migration in migrations.into_iter() {
         if let Target::Version(input_target) = target {
-            if input_target < migration.version {
+            if (input_target as i32) < migration.version {
                 break;
             }
         }
